third_party/WebKit/LayoutTests/inspector/sources/debugger/promise-pane.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>
var p1, p2, p3;

function testFunction()
{
    debugger; // <- will turn on async call stacks here.
    setTimeout(func1, 0);
}

function func1()
{
    p1 = Promise.reject(new Error("EXPECTED"));
    setTimeout(func2, 0);
}

function func2()
{
    var resolve;
    var reject;
    p2 = new Promise(
        function (a, b)
        {
            resolve = a;
            reject = b;
        }
    );
    setTimeout(func3.bind(null, resolve, reject), 50);
}

function func3(resolve, reject)
{
    p3 = p1.catch(function() {});
    var x = 42;
    resolve(x);
    setTimeout(func4, 0);
}

function func4()
{
    debugger; // <- will stop the test here.
}

function test()
{
    var maxAsyncCallStackDepth = 4;
    var promisePane;

    InspectorTest.addSniffer(WebInspector.TabbedPane.prototype, "changeTabView", onChangeTabView, true);
    WebInspector.inspectorView.showViewInDrawer("promises", true);

    function onChangeTabView(id, view)
    {
        if (!promisePane && id === "promises") {
            promisePane = view;
            InspectorTest.assertTrue(promisePane instanceof WebInspector.PromisePane);
            InspectorTest.startDebuggerTest(step1, true);
        }
    }

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2()
    {
        DebuggerAgent.setAsyncCallStackDepth(maxAsyncCallStackDepth, step3);
    }

    function step3()
    {
        InspectorTest.addResult("Is drawer visible: " + WebInspector.inspectorView.drawerVisible());
        InspectorTest.addResult("Selected view in drawer: " + WebInspector.inspectorView.selectedViewInDrawer());
        InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(InspectorTest, step4));
    }

    function step4()
    {
        InspectorTest.addSniffer(promisePane._throttler, "_processCompletedForTests", step5);
    }

    function step5()
    {
        var result = InspectorTest.dumpDataGridIntoString(promisePane._dataGrid);
        result = result.replace(/\|\s+\d+\s+\w*\s*\|/g, "| <number> ms |");
        InspectorTest.addResult(result);
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests promise pane view.
</p>
</body>
</html>
