third_party/WebKit/LayoutTests/inspector/sources/debugger/promise-pane-expected.txt

CONSOLE ERROR: line 16: Uncaught (in promise) Error: EXPECTED
Tests promise pane view.

Set timer for test function.
Is drawer visible: true
Selected view in drawer: promises
Rejected  | func1   | promise-pane.html:16 | promise-pane.html:16 | <number> ms | |
Fulfilled |   func3 | promise-pane.html:36 |                      | <number> ms | |
Fulfilled | func2   | promise-pane.html:24 | promise-pane.html:38 | <number> ms | |
